Add generation of UDID
